Add files via upload
This analysis was carried out because I wanted to find out the factors that cause stress in the lives of students. I wanted to understand how factors such as blood pressure, anxiety, living conditions, basic needs, bullying, and how student-teacher relationship could influence a student's stress levels. This project contains an in-depth look at these factors, how they promote stress in students, and how they each interact with one another to influence the academic performance of the students and their future prospects. At the end of the analysis, recommendations are highlighted to guide stakeholders on how to make decisions based on the analysis. This dataset was gotten from kaggle.com